Add conf_dir attribute.
In order to customize php when running under fpm, you need to be able to
change the configuration settings.  On ubuntu, that means adding files
in /etc/php5/fpm/conf.d.  This cookbook should at least expose that
location so that files can be added there without having to switch on
distro in every other cookbook.